Renamed Timeline component
